requirements.txt
openai==0.28
qdrant-client
python-dotenv